pyproject.toml
[project]
name = "claim-agent"
version = "0.1.0"
description = "Agentic Claim Representative POC using CrewAI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "crewai>=0.121.0",
    "crewai-tools>=0.40.0",
    "mcp>=1.9.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[project.scripts]
claim-agent = "claim_agent.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/claim_agent"]

[tool.hatch.build.targets.sdist]
include = ["/src/claim_agent"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
